move determine account tests to main_test.py
