+++ certbot/tests/main_test.py
@@ -9,6 +9,7 @@ import pytz
 import mock
 
 from certbot import cli
+from certbot import account
 from certbot import colored_logging
 from certbot import constants
 from certbot import configuration
@@ -254,6 +255,70 @@ class MakeOrVerifyCoreDirTest(unittest.TestCase):
         mock_make_or_verify.side_effect = OSError
         self.assertRaises(errors.Error, self._call,
                           self.dir, 0o700, os.geteuid(), False)
+
+
+class DetermineAccountTest(unittest.TestCase):
+    """Tests for certbot.main._determine_account."""
+
+    def setUp(self):
+        self.args = mock.MagicMock(account=None, email=None,
+            register_unsafely_without_email=False)
+        self.config = configuration.NamespaceConfig(self.args)
+        self.accs = [mock.MagicMock(id='x'), mock.MagicMock(id='y')]
+        self.account_storage = account.AccountMemoryStorage()
+
+    def _call(self):
+        # pylint: disable=protected-access
+        from certbot.main import _determine_account
+        with mock.patch('certbot.main.account.AccountFileStorage') as mock_storage:
+            mock_storage.return_value = self.account_storage
+            return _determine_account(self.config)
+
+    def test_args_account_set(self):
+        self.account_storage.save(self.accs[1])
+        self.config.account = self.accs[1].id
+        self.assertEqual((self.accs[1], None), self._call())
+        self.assertEqual(self.accs[1].id, self.config.account)
+        self.assertTrue(self.config.email is None)
+
+    def test_single_account(self):
+        self.account_storage.save(self.accs[0])
+        self.assertEqual((self.accs[0], None), self._call())
+        self.assertEqual(self.accs[0].id, self.config.account)
+        self.assertTrue(self.config.email is None)
+
+    @mock.patch('certbot.client.display_ops.choose_account')
+    def test_multiple_accounts(self, mock_choose_accounts):
+        for acc in self.accs:
+            self.account_storage.save(acc)
+        mock_choose_accounts.return_value = self.accs[1]
+        self.assertEqual((self.accs[1], None), self._call())
+        self.assertEqual(
+            set(mock_choose_accounts.call_args[0][0]), set(self.accs))
+        self.assertEqual(self.accs[1].id, self.config.account)
+        self.assertTrue(self.config.email is None)
+
+    @mock.patch('certbot.client.display_ops.get_email')
+    def test_no_accounts_no_email(self, mock_get_email):
+        mock_get_email.return_value = 'foo@bar.baz'
+
+        with mock.patch('certbot.main.client') as client:
+            client.register.return_value = (
+                self.accs[0], mock.sentinel.acme)
+            self.assertEqual((self.accs[0], mock.sentinel.acme), self._call())
+        client.register.assert_called_once_with(
+            self.config, self.account_storage, tos_cb=mock.ANY)
+
+        self.assertEqual(self.accs[0].id, self.config.account)
+        self.assertEqual('foo@bar.baz', self.config.email)
+
+    def test_no_accounts_email(self):
+        self.config.email = 'other email'
+        with mock.patch('certbot.main.client') as client:
+            client.register.return_value = (self.accs[1], mock.sentinel.acme)
+            self._call()
+        self.assertEqual(self.accs[1].id, self.config.account)
+        self.assertEqual('other email', self.config.email)
 
 
 if __name__ == '__main__':
